footballp_tracker-MinhNK7/pyproject.toml
[project]
name = "footballdetect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "easydict>=1.10",
    "faiss-cpu>=1.7.4",
    "filterpy>=1.4.5",
    "hydra-core===1.3.2",
    "hydra-optuna-sweeper==1.2.0",
    "kornia>=0.8.1",
    "lapx>=0.5.12",
    "lsq-ellipse==2.2.1",
    "matplotlib>=3.7.2",
    "notebook==6.5.3",
    "numpy>=1.23.0",
    "opencv-contrib-python>=4.8.0",
    "opencv-python>=4.8.0",
    "optuna<3.0.0",
    "pandas>=2.0.3",
    "pillow>=9.5.0",
    "psutil>=5.9.5",
    "pynvvideocodec>=2.0.2",
    "pytorch-argus==1.0.0",
    "pyyaml>=6.0.1",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "soccernet==0.1.51",
    "supervision>=0.25.0",
    "tabulate>=0.9.0",
    "tensorboard>=2.20.0",
    "termcolor>=2.3.0",
    "timm>=1.0.21",
    "torch==2.0.0",
    "torchvision==0.15.1",
    "tqdm>=4.66.0",
    "ultralytics>=8.0.100",
    "yacs>=0.1.8",
]

[tool.uv.sources]
torch = [
    { index = "pytorch" }
]
torchvision = [
    { index = "pytorch" }
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
explicit = true
